+++ packages/core/tests/asset.test.ts
@@ -1,6 +1,6 @@
 import { FONT_EXTENSIONS } from '@rsbuild/shared';
 import { createStubRsbuild } from '@scripts/test-helper';
-import { pluginAsset, getRegExpForExts } from '../../src/plugins/asset';
+import { pluginAsset, getRegExpForExts } from '../src/plugins/asset';
 
 describe('getRegExpForExts', () => {
   it('should get correct RegExp of exts', () => {
