+++ packages/core/tests/default.test.ts
@@ -1,6 +1,6 @@
 import { createStubRsbuild } from '@scripts/test-helper';
 import type { RsbuildPlugin } from '@rsbuild/shared';
-import { BUILTIN_LOADER } from '../../src/provider/shared';
+import { BUILTIN_LOADER } from '../src/provider/shared';
 
 describe('applyDefaultPlugins', () => {
   it('should apply default plugins correctly', async () => {
