+++ packages/core/tests/define.test.ts
@@ -1,5 +1,5 @@
 import { createStubRsbuild } from '@scripts/test-helper';
-import { pluginDefine } from '../../src/plugins/define';
+import { pluginDefine } from '../src/plugins/define';
 
 describe('plugin-define', () => {
   test('should register define plugin correctly', async (item) => {
