+++ packages/core/tests/html.test.ts
@@ -1,55 +1,277 @@
-import { hasTitle } from '../src/rspack/HtmlBasicPlugin';
-import { getHTMLPathByEntry } from '../src/provider/core/initPlugins';
-import type { NormalizedConfig } from '@rsbuild/shared';
+import { pluginHtml } from '../src/plugins/html';
+import { pluginEntry } from '../src/plugins/entry';
+import { createStubRsbuild } from '@scripts/test-helper';
+import type { HtmlConfig } from '@rsbuild/shared';
 
-test('should detect HTML title via "hasTitle" correctly', () => {
-  expect(hasTitle()).toEqual(false);
-  expect(hasTitle('<title></title>')).toEqual(true);
-  expect(hasTitle('<title>hello</title>')).toEqual(true);
-  expect(hasTitle('<title  ></title>')).toEqual(true);
-  expect(hasTitle('<title></title  >')).toEqual(true);
-  expect(hasTitle('<title >hello</title  >')).toEqual(true);
-  expect(hasTitle('<html><head></head><body></body></html>')).toEqual(false);
-  expect(
-    hasTitle(
-      '<html><head><title>Page Title</title></head><body></body></html>',
-    ),
-  ).toEqual(true);
-  expect(
-    hasTitle(
-      '<html>\n<head>\n<title>\nPage Title\n</title>\n</head>\n<body></body>\n</html>',
-    ),
-  ).toEqual(true);
+vi.mock('@rsbuild/shared', async (importOriginal) => {
+  const mod = await importOriginal<any>();
+  return {
+    ...mod,
+    isFileExists: async () => true,
+  };
 });
 
-describe('getHTMLPathByEntry', () => {
-  it('should use distPath.html as the folder', async () => {
-    const htmlPath = getHTMLPathByEntry('main', {
-      output: {
-        distPath: {
-          html: 'my-html',
-        },
-      },
-      html: {
-        outputStructure: 'nested',
-      },
-    } as NormalizedConfig);
+describe('plugin-html', () => {
+  it('should register html plugin correctly', async () => {
+    const rsbuild = await createStubRsbuild({
+      plugins: [pluginEntry(), pluginHtml()],
+    });
+    const config = await rsbuild.unwrapConfig();
 
-    expect(htmlPath).toEqual('my-html/main/index.html');
+    expect(await rsbuild.matchBundlerPlugin('HtmlWebpackPlugin')).toBeTruthy();
+    expect(config).toMatchSnapshot();
   });
 
-  it('should allow to disable html folder', async () => {
-    const htmlPath = getHTMLPathByEntry('main', {
-      output: {
-        distPath: {
-          html: 'html',
+  it('should not register html plugin when target is node', async () => {
+    const rsbuild = await createStubRsbuild({
+      plugins: [pluginEntry(), pluginHtml()],
+      rsbuildConfig: {
+        output: {
+          targets: ['node'],
         },
       },
-      html: {
-        outputStructure: 'flat',
-      },
-    } as NormalizedConfig);
-
-    expect(htmlPath).toEqual('html/main.html');
+    });
+    expect(await rsbuild.matchBundlerPlugin('HtmlWebpackPlugin')).toBeFalsy();
   });
+
+  it('should not register html plugin when target is web-worker', async () => {
+    const rsbuild = await createStubRsbuild({
+      plugins: [pluginEntry(), pluginHtml()],
+      rsbuildConfig: {
+        output: {
+          targets: ['web-worker'],
+        },
+      },
+    });
+    expect(await rsbuild.matchBundlerPlugin('HtmlWebpackPlugin')).toBeFalsy();
+  });
+
+  it('should not register html plugin when target is service-worker', async () => {
+    const rsbuild = await createStubRsbuild({
+      plugins: [pluginEntry(), pluginHtml()],
+      rsbuildConfig: {
+        output: {
+          targets: ['service-worker'],
+        },
+      },
+    });
+    expect(await rsbuild.matchBundlerPlugin('HtmlWebpackPlugin')).toBeFalsy();
+  });
+
+  it('should register nonce plugin when using security.nonce', async () => {
+    const rsbuild = await createStubRsbuild({
+      plugins: [pluginEntry(), pluginHtml()],
+      rsbuildConfig: {
+        security: {
+          nonce: 'test-nonce',
+        },
+      },
+    });
+
+    expect(await rsbuild.matchBundlerPlugin('HtmlNoncePlugin')).toBeTruthy();
+  });
+
+  it('should register crossorigin plugin when using html.crossorigin', async () => {
+    const rsbuild = await createStubRsbuild({
+      plugins: [pluginEntry(), pluginHtml()],
+      rsbuildConfig: {
+        html: {
+          crossorigin: true,
+        },
+      },
+    });
+
+    const config = await rsbuild.unwrapConfig();
+    expect(
+      await rsbuild.matchBundlerPlugin('HtmlCrossOriginPlugin'),
+    ).toBeTruthy();
+    expect(config.output?.crossOriginLoading).toEqual('anonymous');
+  });
+
+  it('should register appIcon plugin when using html.appIcon', async () => {
+    const rsbuild = await createStubRsbuild({
+      plugins: [pluginEntry(), pluginHtml()],
+      rsbuildConfig: {
+        html: {
+          appIcon: './src/assets/icon.png',
+        },
+      },
+    });
+
+    expect(await rsbuild.matchBundlerPlugin('HtmlAppIconPlugin')).toBeTruthy();
+  });
+
+  it('should allow to set favicon by html.favicon option', async () => {
+    const rsbuild = await createStubRsbuild({
+      plugins: [pluginEntry(), pluginHtml()],
+      rsbuildConfig: {
+        html: {
+          favicon: './src/favicon.ico',
+        },
+      },
+    });
+    const config = await rsbuild.unwrapConfig();
+
+    expect(config).toMatchSnapshot();
+  });
+
+  it('should allow to set inject by html.inject option', async () => {
+    const rsbuild = await createStubRsbuild({
+      plugins: [pluginEntry(), pluginHtml()],
+      rsbuildConfig: {
+        html: {
+          inject: 'body',
+        },
+      },
+    });
+    const config = await rsbuild.unwrapConfig();
+
+    expect(config).toMatchSnapshot();
+  });
+
+  it('should enable minify in production', async () => {
+    const { NODE_ENV } = process.env;
+    process.env.NODE_ENV = 'production';
+
+    const rsbuild = await createStubRsbuild({
+      plugins: [pluginEntry(), pluginHtml()],
+    });
+    const config = await rsbuild.unwrapConfig();
+
+    expect(config).toMatchSnapshot();
+
+    process.env.NODE_ENV = NODE_ENV;
+  });
+
+  it('should allow to modify plugin options by tools.htmlPlugin', async () => {
+    const rsbuild = await createStubRsbuild({
+      plugins: [pluginEntry(), pluginHtml()],
+      rsbuildConfig: {
+        tools: {
+          htmlPlugin(_config, utils) {
+            expect(utils.entryName).toEqual('index');
+            return {
+              inject: true,
+            };
+          },
+        },
+      },
+    });
+    const config = await rsbuild.unwrapConfig();
+
+    expect(config).toMatchSnapshot();
+  });
+
+  it('should allow to disable html plugin', async () => {
+    const rsbuild = await createStubRsbuild({
+      plugins: [pluginEntry(), pluginHtml()],
+      rsbuildConfig: {
+        tools: {
+          htmlPlugin: false,
+        },
+      },
+    });
+
+    expect(await rsbuild.matchBundlerPlugin('HtmlWebpackPlugin')).toBeFalsy();
+  });
+
+  it('should disable html plugin when htmlPlugin is an array and contains false', async () => {
+    const rsbuild = await createStubRsbuild({
+      plugins: [pluginEntry(), pluginHtml()],
+      rsbuildConfig: {
+        tools: {
+          htmlPlugin: [{}, false],
+        },
+      },
+    });
+
+    expect(await rsbuild.matchBundlerPlugin('HtmlWebpackPlugin')).toBeFalsy();
+  });
+
+  it('should support multi entry', async () => {
+    const rsbuild = await createStubRsbuild({
+      plugins: [pluginEntry(), pluginHtml()],
+      rsbuildConfig: {
+        source: {
+          entry: {
+            main: './src/main.ts',
+            foo: './src/foo.ts',
+          },
+        },
+        html: {
+          template({ entryName }) {
+            expect(['main', 'foo'].includes(entryName)).toBeTruthy();
+          },
+        },
+      },
+    });
+    const config = await rsbuild.unwrapConfig();
+
+    expect(config).toMatchSnapshot();
+  });
+
+  it('should add one tags plugin instance', async () => {
+    const rsbuild = await createStubRsbuild({
+      plugins: [pluginEntry(), pluginHtml()],
+      rsbuildConfig: {
+        source: {
+          entry: {
+            main: './src/main.ts',
+            foo: './src/foo.ts',
+          },
+        },
+        html: {
+          tags: { tag: 'script', attrs: { src: 'jq.js' } },
+          tagsByEntries: {},
+        },
+      },
+    });
+    const config = await rsbuild.unwrapConfig();
+    const plugins = config.plugins?.filter(
+      (p: { name: string }) => p.name === 'HtmlTagsPlugin',
+    );
+    expect(plugins?.length).toBe(1);
+    expect(config).toMatchSnapshot();
+  });
+
+  it('should add tags plugin instances for each entries', async () => {
+    const rsbuild = await createStubRsbuild({
+      plugins: [pluginEntry(), pluginHtml()],
+      rsbuildConfig: {
+        source: {
+          entry: {
+            main: './src/main.ts',
+            foo: './src/foo.ts',
+          },
+        },
+        html: {
+          tags: [{ tag: 'script', attrs: { src: 'jq.js' } }],
+          tagsByEntries: {
+            foo: [{ tag: 'script', attrs: { src: 'foo.js' } }],
+          },
+        },
+      },
+    });
+    const config = await rsbuild.unwrapConfig();
+    const plugins = config.plugins?.filter(
+      (p: { name: string }) => p.name === 'HtmlTagsPlugin',
+    );
+    expect(plugins?.length).toBe(2);
+    expect(config).toMatchSnapshot();
+  });
+
+  it.each<{ value: HtmlConfig['inject']; message: string }>([
+    { value: false, message: 'false' },
+    { value: () => false, message: '() => false' },
+  ])(
+    'should stop injecting <script> if inject is $message',
+    async ({ value }) => {
+      const rsbuild = await createStubRsbuild({
+        plugins: [pluginEntry(), pluginHtml()],
+        rsbuildConfig: { html: { inject: value } },
+      });
+      const config = await rsbuild.unwrapConfig();
+      expect(config).toMatchSnapshot();
+    },
+  );
 });
